+++ backend/canisters/community/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [unreleased]
 
+### Fixed
+
+- Fix check for if user can leave community ([#7783](https://github.com/open-chat-labs/open-chat/pull/7783))
+
 ## [[2.0.1699](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.1699-community)] - 2025-04-10
 
 ### Changed